+++ completions/kkfetch.bash
@@ -13,7 +13,7 @@ _kkfetch() {
         prev="$3"
     fi
 
-    local opts="-m -d -l -h -V --modules --disable --no-color --logo --no-logo --list-modules --disk-path --json --help --version"
+    local opts="-m -d -l -h -V --modules --disable --no-color --logo --no-logo --list-modules --disk-path --json --timings --no-plugins --doctor --help --version"
     local modules="title os host kernel installed uptime packages pkgs shell display desktop de wm terminal term cpu gpu memory mem swap disk battery localip theme icons colors palette"
     local logos="ferris rust debian ubuntu linuxmint mint fedora arch archlinux rhel redhat centos rocky rockylinux almalinux alma endeavouros endeavour manjaro generic tux linux opensuse suse alpine gentoo void pop popos nixos kali freebsd slackware artix zorin none"